Move IdentityHelper & tests to common
