+++ src/System.IO.IsolatedStorage/src/System/IO/IsolatedStorage/Helper.Win32.Unix.cs
@@ -4,6 +4,7 @@
 
 using System.Reflection;
 using System.Runtime.InteropServices;
+using System.Security;
 using System.Security.AccessControl;
 using System.Security.Principal;
 
@@ -125,7 +126,7 @@ namespace System.IO.IsolatedStorage
             AssemblyName assemblyName = assembly.GetName();
             Uri codeBase = new Uri(assembly.CodeBase);
 
-            hash = GetNormalizedStrongNameHash(assemblyName);
+            hash = IdentityHelper.GetNormalizedStrongNameHash(assemblyName);
             if (hash != null)
             {
                 hash = "StrongName" + separator + hash;
@@ -133,7 +134,7 @@ namespace System.IO.IsolatedStorage
             }
             else
             {
-                hash = GetNormalizedUriHash(codeBase);
+                hash = IdentityHelper.GetNormalizedUriHash(codeBase);
                 hash = "Url" + separator + hash;
                 identity = codeBase;
             }
